Move /set into own folder and refactor
